requirements.txt
streamlit>=1.32.0
edge-tts>=6.1.3
google-generativeai>=0.3.0
pydub>=0.25.1